This example works on OS X and possibly on Linux (untested).
All examples I've tried to follow depended on conio.h for asyncronous
input, but conio.h is a MS-DOS header and is not included on OS X or
Linux. Also this example is the prettiest of all.

Thanks to René Hansen for sharing his knowledge, please do read more
here:
https://rhardih.io/2010/04/using-ncurses-in-c-for-text-user-interfaces-featuring-xcode/